Compute residual scale for SVGImage drawing using the integer size
The (rounded) integer size is used when sizing the underlying
Frame(View), so we need to use it, rather than the LayoutSize, to
compute the residual scale to avoid having the image 'spill' out of the
viewport.

This also improves the case in crbug.com/1123047 enough to call it
fixed. Add a test for it.

Bug: 1212634, 1163178, 1123047
Change-Id: I9ea8a7ae99ae195c755db036686f51a06f6ac59e