Add |SideEffectsDisabled| for MinMax-after-layout
Computing MinMax after layout was prohibited in legacy. NG can
do this, except that it copies the fragment data back to the
|LayoutObject| tree, update |PaintLayer|, scroll bar
propoerties, etc.

This patch adds |SideEffectsDisabled| to indicate the
situation, and change |NGBlockNode::FinishLayout| to not to
update the |LayoutObject| tree and other global states.

This is the case 1 of the discussion doc:
https://docs.google.com/document/d/1HkMwnfr-JbP2SE0ktYalkeAutwciuy-GftLs3TOnBcI/edit?usp=sharing

Bug: 1213358
Change-Id: I352d73447e1795510f3673abb8c2dc532afea532